requirements.txt
python-telegram-bot==20.3
openai
pytz
apscheduler
python-dotenv
urllib3
six
nest_asyncio
gspread
oauth2client
spotipy
requests
telegram
